pyproject.toml
[project]
name = "reversible"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pillow>=11.1.0",
    "sqlmodel>=0.0.23",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pydantic-settings>=2.8.1",
    "pyfakefs>=5.7.4",
    "pyright>=1.1.396",
    "pytest>=8.3.5",
]
